Move getPepos to prompts/repos
